Crate webmasters3

Features

Handle the following Resources with ease from the central hub ...

sitemaps (delete, get, list and submit)

sites (add, delete, get and list)

urlcrawlerrorscounts (query)

urlcrawlerrorssamples (get, list and markasfixed)

Everything else about the webmasters API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.sitemaps().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate webmasters3;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Webmasters::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.sitemaps().get("siteUrl", "feedpath")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

SiteAddMethodBuilder
Adds a site to the set of the user's sites in Webmaster Tools.

SiteDeleteMethodBuilder
Removes a site from the set of the user's Webmaster Tools sites.

SiteGetMethodBuilder
Retrieves information about specific site.

SiteListMethodBuilder
Lists your Webmaster Tools sites.

SiteMethodsBuilder
A builder providing access to all methods supported on site resources. It is not used directly, but through the Webmasters hub.

SitemapDeleteMethodBuilder
Deletes a sitemap from this site.

SitemapGetMethodBuilder
Retrieves information about a specific sitemap.

SitemapListMethodBuilder
Lists sitemaps uploaded to the site.

SitemapMethodsBuilder
A builder providing access to all methods supported on sitemap resources. It is not used directly, but through the Webmasters hub.

SitemapSubmitMethodBuilder
Submits a sitemap for a site.

SitemapsListResponse
List of sitemaps.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

SitesListResponse
List of sites with access level information.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

UrlCrawlErrorCount
An entry in a URL crawl errors time series.

This type is not used in any activity, and only used as part of another schema.

UrlCrawlErrorCountsPerType
Number of errors per day for a specific error type (defined by platform and category).

This type is not used in any activity, and only used as part of another schema.

UrlCrawlErrorsCountsQueryResponse
A time series of the number of URL crawl errors per error category and platform.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query (response)

UrlCrawlErrorsSample
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

UrlCrawlErrorsSamplesListResponse
List of crawl error samples.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

UrlSampleDetails
Additional details about the URL, set only when calling get().

This type is not used in any activity, and only used as part of another schema.

UrlcrawlerrorscountMethodsBuilder
A builder providing access to all methods supported on urlcrawlerrorscount resources. It is not used directly, but through the Webmasters hub.

UrlcrawlerrorscountQueryMethodBuilder
Retrieves a time series of the number of URL crawl errors per error category and platform.

UrlcrawlerrorssampleGetMethodBuilder
Retrieves details about crawl errors for a site's sample URL.

UrlcrawlerrorssampleListMethodBuilder
Lists a site's sample URLs for the specified crawl error category and platform.

UrlcrawlerrorssampleMarkAsFixedMethodBuilder
Marks the provided site's sample URL as fixed, and removes it from the samples list.

UrlcrawlerrorssampleMethodsBuilder
A builder providing access to all methods supported on urlcrawlerrorssample resources. It is not used directly, but through the Webmasters hub.

Webmasters
Central instance to access all Webmasters related resource activities

WmxSite
Access level information for a Webmaster Tools site.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

WmxSitemap
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

WmxSitemapContent
Information about the various content types in the sitemap.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.